♻️ Extract parse functions
