+++ src/url.rs
@@ -4,7 +4,8 @@
 
 use scraper::Html;
 use url::Url;
-use crate::{APP_USER_AGENT, APPLE_DEV_RELEASES, GenericError, GenericResult, parse_article_title, SELECTORS};
+use crate::{APP_USER_AGENT, APPLE_DEV_RELEASES, GenericError, GenericResult, SELECTORS};
+use crate::parse::parse_article_title;
 
 /// Gets a URL and returns the body of the response.
 ///
